Add files via upload
Sarah has just learned to type and access the internet. As soon as he entered the Internet, she decided to enter a chat room and say hello to everyone. Sara entered a word in the chat room. If it is possible to delete some of the letters of the word that Sarah entered and the word hello remains at the end, it means that Sarah was able to say hello, otherwise, no.


It is guaranteed that the input only consists of English lowercase letters.